Create generic polymorphic converter
