+++ src/Mollie.Api/Framework/Factories/ITypeFactory.cs
@@ -1,0 +1,5 @@
+namespace Mollie.Api.Framework.Factories;
+
+internal interface ITypeFactory<T> {
+    T Create(string? type);
+}
